Add type option to parameter fields
Add the ability to specify a `swagger:parameters` struct field's type
when generating a spec.

For example:

```
type MyString struct {
  Language language.Tag
}

// swagger:parameters getPost
type Post struct {
  // in: query
  // type: string
  Title MyString `json:"title"`
}
```

This allows the API to convert the string into a MyString struct, and pass
the actual Post struct around. Currently, one would have to specify that
Title is a string, and then create another struct in which Title is a
MyString.